requirements.txt
asgiref==3.3.4
bower==0.0.0
Django==3.2.4
django-js-asset==1.2.2
django-tinymce==3.3.0
django-wangeditor==1.0.1
DjangoUeditor==1.8.143
Jinja2==3.0.1
MarkupSafe==2.0.1
pytz==2021.1
shortid8==1.0
sqlparse==0.4.1
tiny==0.1.3

